requirements.txt
click>=8.0
alpaca-py>=0.21.0
python-dotenv>=1.0
rich>=13.0
